backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic-settings==2.8.1
python-multipart==0.0.20
pytest==8.3.5
httpx==0.28.1
oss2==2.19.1
sqlalchemy==2.0.40
PyMySQL==1.1.1
Pillow==11.2.1
